DEVELOPMENT OF SMART COOLING SYSTEM FOR STAND-ALONE PHOTOVOLTAIC
Project aims to design a smart cooling system for photovoltaic solar panels that utilizes IoT technology through Cayenne to monitor and control the flow of cooling water and the operation of fans. The system will use data collected from the solar panel's temperature and efficiency to optimize performance and power production. Additionally, the project will investigate the amount of cooling water required to achieve the highest photovoltaic efficiency in the solar cell by studying the relationship between the cooling water flow rate, temperature and the efficiency of the solar cell. The goal is to improve the overall efficiency and power production of the solar panel system through real-time monitoring and control via IoT technology.